% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{autoselect.mix.sd}
\alias{autoselect.mix.sd}
\title{autoselect.mix.sd}
\usage{
autoselect.mix.sd(betahat, sebetahat, mult)
}
\arguments{
\item{betahat}{is an J x R matrix of betahats (or t statistics)}

\item{sebetahat}{is a J x R matrix of their standard errors (or 1s, if T staitsitcs)}
}
\value{
mult^((-npoint):0) * sigmaamax)
}

