% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testscript.R
\name{autoselect.omega}
\alias{autoselect.omega}
\title{autoselect.omega}
\usage{
autoselect.omega(betahat, sebetahat, minparam, mult, maxparam)
}
\arguments{
\item{betahat}{= column of betahats (or zs)}

\item{minparam}{= what the standard error is divided by (want to be small)}

\item{mult}{= density parameter (intervals between points)}

\item{maxparam}{= how big the grid needs to be}

\item{sebetahat=}{column of standard errs}
}
\value{
vector of variances for one tissue
}
\description{
autoselect.omega
}
\concept{
this computes a rough largest value you'd want to use, based on idea that sigmaamax^2 + sebetahat^2 should be at least betahat^2
}

