% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationswithmean.R
\name{chat_sim}
\alias{chat_sim}
\title{chat_sim}
\usage{
chat_sim(n = 1000, d = 3, betasd = 1, esd = 0.1, K = 10)
}
\arguments{
\item{n}{number of SNPs}

\item{d}{number of subgroups}

\item{betasd}{the vriance of true effect}

\item{esd}{the standard error of E in chat=mu+c+E, i.e., E~N(0,diag(esd^2))}
}

