% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationswithmean.R
\name{chat_sim_fsimple}
\alias{chat_sim_fsimple}
\title{chat_sim_fsimple}
\usage{
chat_sim_fsimple(n = 1000, d = 8, betasd = 1, esd = 0.1, K = 10)
}
\arguments{
\item{n}{number of SNPs}

\item{d}{number of subgroups}

\item{betasd}{the vriance of true effect}

\item{esd}{the standard error of E in chat=mu+c+E, i.e., E~N(0,diag(esd^2))}
}
\value{
omega the size of the effects

f the factors from which they were simulated
}

