% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{compare.func}
\alias{compare.func}
\title{compares upper and lower tail probability at each tissue to determine larger}
\usage{
compare.func(j)
}
\arguments{
\item{all.upper}{and all.lower are JxR matrices of upper and lower tail probabilities for all gene pairs}
}
\value{
1 x R vector of lfsr at each tissue
}
\description{
compares upper and lower tail probability at each tissue to determine larger
}

