% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.covmat.bmafull.only}
\alias{compute.covmat.bmafull.only}
\title{compute.covmat.bmafull.only}
\usage{
compute.covmat.bmafull.only(b.gp.hat, sebetahat, A)
}
\arguments{
\item{b.gp.hat}{a JxR matrix of betahats}

\item{sebetahat}{a JxR matrix of their standard errors}
}
\value{
A list of covariance matrices
}
