% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.covmat.nopc}
\alias{compute.covmat.nopc}
\title{get.prior.covar.nopcs}
\usage{
compute.covmat.nopc(b.gp.hat, sebetahat, Q, t.stat, lambda.mat, A, factor.mat,
  bma = TRUE)
}
\description{
get.prior.covar.nopcs
}

