% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthscripts.R
\name{compute.covmat.using.fixed.omega.withQ}
\alias{compute.covmat.using.fixed.omega.withQ}
\title{compute.covmat.using.fixed.omega.withQ}
\usage{
compute.covmat.using.fixed.omega.withQ(max.step, Q, bma = TRUE, omega = 2,
  zero = TRUE, A, power = 1)
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles}

\item{se.gp.hat}{JxR matrix of standarderrors}
}

