% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthscripts.R
\name{compute.covmat.using.truth.fixed.omega}
\alias{compute.covmat.using.truth.fixed.omega}
\title{compute.covmat.using.truth.fixed.omega}
\usage{
compute.covmat.using.truth.fixed.omega(covlist, A, omega, zero = TRUE)
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles}

\item{se.gp.hat}{JxR matrix of standarderrors}
}
