% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthscripts.R
\name{compute.covmat.using.truth}
\alias{compute.covmat.using.truth}
\title{compute.covmat.using.truth}
\usage{
compute.covmat.using.truth(b.gp.hat, sebetahat, A, zero = FALSE, covlist,
  maxp = 1, minp = 1, power = 1)
}
\arguments{
\item{b.gp.hat}{JxR matrix of mles}

\item{covmat}{a K list of covariance matrices}

\item{se.gp.hat}{JxR matrix of standarderrors}
}

