% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.covmat.with.heterogeneity.no.shrink}
\alias{compute.covmat.with.heterogeneity.no.shrink}
\title{compute.covmat.with.heterogeneity.no.shrink}
\usage{
compute.covmat.with.heterogeneity.no.shrink(b.gp.hat, sebetahat, A,
  zero = FALSE)
}
\arguments{
\item{b.gp.hat}{a JxR matrix of betahats}

\item{sebetahat}{a JxR matrix of their standard errors}
}
\value{
A list of covariance matrices
}
\description{
compute.covmat.with.heterogeneity.no.shrink
}

