% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.covmat.with.rho}
\alias{compute.covmat.with.rho}
\title{compute.covmat.with.rho}
\usage{
compute.covmat.with.rho(b.gp.hat, sebetahat, Q, lambda.mat, P, A, factor.mat,
  t.stat)
}
\description{
compute.covmat.with.rho
}

