% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testscript.R
\name{compute.hm.covmat.all.max.step.omega}
\alias{compute.hm.covmat.all.max.step.omega}
\title{compute.hm.covmat.all.max.step.omega}
\usage{
compute.hm.covmat.all.max.step.omega(mult = sqrt(2), minparam = 4, b.hat,
  se.hat, t.stat, v.j, Q, lambda.mat, A, factor.mat, max.step, multone,
  maxparam)
}
\arguments{
\item{minparam}{= what the standard error is divided by (want to be small)}

\item{t.stat}{matrix of strong t statistics (needs to be R in columns but not necessarily j in rows)}

\item{Q}{number of factors to consider}

\item{max.step}{= object with denoised covariance matrices}

\item{multone}{= density parameter (intervals between points for within tissue variances, default 2 so that the ratio between sd is sqrt(2))}

\item{maxparam}{= how big the grid needs to be}

\item{betahat}{= JxR matrix of betahats (or zs)}

\item{sebetahat=}{JxR matrix of standard errs}

\item{factor}{mat a fxR matrix of factors}

\item{BMA=true}{wheter or not to use singleton and shared config}
}
\value{
A list of covariance matrices
}
\description{
compute.hm.covmat.all.max.step.omega
}

