% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{compute.hm.covmat.all.max.step}
\alias{compute.hm.covmat.all.max.step}
\title{compute.hm.covmat.all.max.step}
\usage{
compute.hm.covmat.all.max.step(b.hat, se.hat, t.stat, v.j, Q, lambda.mat, A,
  factor.mat, max.step, zero = FALSE)
}
\arguments{
\item{b.hat,se.hat}{matrices to help choose grid weights (JxR)}

\item{v.j}{a JxR matrix of standard errors for scaling (in the zstat case this is all 1s)}

\item{Q}{number of single rank factors to include in the set of covariance matrices}

\item{max.step:}{the list ouput of deconvolution EM (list of length 2, contains the denoised covariance  matrices and vector of pi)}

\item{P}{PC approximation}

\item{BMA}{Whether or not to include singleton and full configurations}
}
\value{
a list of covariance matrices
}
\description{
compute.hm.covmat.all.max.step
}
\details{
use a prespecified X`X from the output of the 'deconvoluting EM'
}

