% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{compute.hm.covmat.all.max.step.withQ}
\alias{compute.hm.covmat.all.max.step.withQ}
\title{compute.hm.covmat.all.max.step.withQ}
\usage{
compute.hm.covmat.all.max.step.withQ(b.hat, se.hat, t.stat, v.j, lambda.mat, A,
  factor.mat, max.step, Q, power = 1, zero = TRUE, maxp = 1, minp = 1)
}
\arguments{
\item{Q}{rank of SFA approximation}

\item{t}{stat}

\item{P}{rank of PC approxiatmion}

\item{permsnp}{= number of strong stats you want to train on (default is 1000)}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\details{
wrapper to compute denoised estimates of the fuller rank covariance matrices
}

