% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{compute.hm.covmat}
\alias{compute.hm.covmat}
\title{compute.hm.covmat}
\usage{
compute.hm.covmat(t.stat, v.j, Q, lambda.mat, P, A, factor.mat, max.step)
}
\arguments{
\item{Q}{number of single rank factors to include in the set of covariance matrices}

\item{P}{PC approximation}

\item{max.step:}{the list ouput of deconvolution EM (list of length 2, contains the denoised covariance  matrices and vector of pi)}

\item{BMA}{Whether or not to include singleton and full configurations}
}
\value{
a list of covariance matrices
}
\details{
use a prespecified X`X from the output of the 'deconvoluting EM'
}
