% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthscripts.R
\name{compute.hm.mash.covmat.all.max.step.fixed.omega}
\alias{compute.hm.mash.covmat.all.max.step.fixed.omega}
\title{compute.hm.mash.covmat.all.max.step.fixed.omega}
\usage{
compute.hm.mash.covmat.all.max.step.fixed.omega(bma = TRUE, omega, Q,
  lambda.mat, A, factor.mat, max.step, zero = FALSE, maxp = 1, minp = 1,
  power = 1)
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles}

\item{se.gp.hat}{JxR matrix of standarderrors}
}
