% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixEm.R
\name{compute.hm.train.log.lik.pen}
\alias{compute.hm.train.log.lik.pen}
\title{compute.hm.train.log.lik.pen}
\usage{
compute.hm.train.log.lik.pen(train.b, se.train, covmat, A, pen)
}
\arguments{
\item{train.b}{=  JxR matrix of training beta hats}

\item{se.train}{= JxR matrix of training standard errors}

\item{covmat}{= LxK dimenstional (unlisted list) of prior covariance matrices}

\item{A}{output file name}

\item{Pen}{likelihood penalty, default1}
}
\value{
An object containing pis and model fit from the EM, and a pdf barplot
}
\description{
= takes a matrix of training sumamry statistics and their standard errors and computes the likelihood matrix according to a list of covariance matrices, using the exponent of the log likelihood - max (llj)
}
