% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixEm.R
\name{compute.hm.train.log.lik.pen.vmat}
\alias{compute.hm.train.log.lik.pen.vmat}
\title{compute.hm.train.log.lik.pen.vmat}
\usage{
compute.hm.train.log.lik.pen.vmat(train.b, covmat, A, pen, train.s, cormat)
}
\arguments{
\item{train.b}{=  JxR matrix of training beta hats}

\item{covmat}{= LxK dimenstional (unlisted list) of prior covariance matrices}

\item{A}{output file name}

\item{train.s}{=  JxR matrix of training serrors}

\item{Pen}{likelihood penalty, default1}

\item{cormatrix}{of correlation of residuals, the same for every gene}
}
\value{
An object containing pis and model fit from the EM, and a pdf barplot
}
\description{
= takes a matrix of training sumamry statistics and their standard errors and computes the likelihood matrix according to a list of covariance matrices, using the exponent of the log likelihood - max (llj)
}

