% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{compute.hm.train.log.lik.pen.with.L}
\alias{compute.hm.train.log.lik.pen.with.L}
\title{compute.hm.train.log.lik.pen.with.L}
\usage{
compute.hm.train.log.lik.pen.with.L(train.b, se.train, covmat, A, pen, L)
}
\arguments{
\item{train.b}{=  JxR matrix of training beta hats (here, already transformed)}

\item{se.train}{= JxR matrix of training standard errors}

\item{covmat}{= LxK dimenstional (unlisted list) of prior covariance matrices}

\item{A}{output file name}

\item{L}{= projection matrix}

\item{Pen}{likelihood penalty, default1}
}
\value{
An object containing pis and model fit from the EM, and a pdf barplot
}
\description{
= takes a matrix of training sumamry statistics and their standard errors and computes the likelihood matrix according to a list of covariance matrices, using the exponent of the log likelihood - max (llj)
}

