% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.lik.test}
\alias{compute.lik.test}
\title{compute.lik.test}
\usage{
compute.lik.test(b.test, J, se.test, covmat, A, pis)
}
\arguments{
\item{b.test}{JxR matrix with the mles for test gene-snp pairs}

\item{J}{number of gene snp pairs to consider}

\item{se.test}{JxR matrix of standard errors of the test set}

\item{covmat}{K list of covariance matrices}

\item{pis}{K matrix of HM weights form compute.hm.train}
}
\description{
compute.lik.test
}

