% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.lik.test.semat}
\alias{compute.lik.test.semat}
\title{compute.lik.test.semat}
\usage{
compute.lik.test.semat(b.test, J, se.test, covmat, A, pis)
}
\arguments{
\item{b.test}{JxR matrix with the mles for test gene-snp pairs}

\item{J}{number of gene snp pairs to consider}

\item{covmat}{K list of covariance matrices}

\item{pis}{K matrix of HM weights form compute.hm.train}

\item{se.mat}{RxR matrix of standard errors of the test set, estimated as the same for all genes}
}
\description{
compute.lik.test.semat
}

