% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{compute.lik.test.withL}
\alias{compute.lik.test.withL}
\title{compute.lik.test.withL}
\usage{
compute.lik.test.withL(b.test, J, se.test, covmat, A, pis, L)
}
\arguments{
\item{b.test}{JxR transformed estimates}

\item{J}{number of gene snp pairs to consider}

\item{se.test}{JxR matrix of standard errors of the test set}

\item{covmat}{K list of covariance matrices}

\item{pis}{K matrix of HM weights form compute.hm.train}

\item{L}{projection matrix}
}

