% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisesharingscripts.R
\name{compute.mag.by.sharing}
\alias{compute.mag.by.sharing}
\title{compute.mag.by.sharing}
\usage{
compute.mag.by.sharing(lfsr.mash, thresh, pm.mash.beta)
}
\details{
computes the matrix of pairwise sharing by magnitude
}

