% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisesharingscripts.R
\name{compute.sharing.by.sig}
\alias{compute.sharing.by.sig}
\title{compute.sharing.by.sig}
\usage{
compute.sharing.by.sig(lfsr.mash, thresh)
}
\details{
computes the matrix of pairwise sharing by significiance
}
