% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisesharingscripts.R
\name{compute.sharing.by.sign}
\alias{compute.sharing.by.sign}
\title{compute.sharing.by.sign}
\usage{
compute.sharing.by.sign(lfsr.mash, thresh, pm.mash.beta)
}
\details{
computes the matrix of pairwise sharing by sign
}
