% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.total.quant}
\alias{compute.total.quant}
\title{compute.total.quant}
\usage{
compute.total.quant(A, J, all.arrays)
}
\arguments{
\item{A}{filename to save weighted quantities in}

\item{J}{number of gene snp pairs to consider}

\item{all.arrays}{list of JxKxR arrays of posterior componenet-specific quantities}

\item{If}{file present, read array file; else use list in memory}
}
