% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{convertstandarderrors}
\alias{convertstandarderrors}
\title{convertstandarderrors}
\usage{
convertstandarderrors(s.j, L)
}
\arguments{
\item{s.j}{matrix of unscaled standard errors}

\item{L}{projection matrix, this can be Rx(R-1)}
}
\value{
the standard errors of wj, diag(LVL)
}
