% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{deconvolution.em.with.bovy.with.L.oneK}
\alias{deconvolution.em.with.bovy.with.L.oneK}
\title{deconvolution.em.with.bovy.with.L.oneK}
\usage{
deconvolution.em.with.bovy.with.L.oneK(t.stat, v.j, L, w)
}
\arguments{
\item{L}{projection matrix, this can be Rx(R-1)}

\item{s.j}{matrix of unscaled standard errors}
}
\value{
the JxRxR (or R-1) array of marginal variances
}
