% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{deconvolution.em.with.bovy.with.L}
\alias{deconvolution.em.with.bovy.with.L}
\title{deconvolution.em.bovy.with.L}
\usage{
deconvolution.em.with.bovy.with.L(t.stat, factor.mat, v.j, lambda.mat, P, L, Q,
  w)
}
\arguments{
\item{t.stat}{the R centered matrix to initalize covariance matrices (should be w plus the additional column)}

\item{P}{rank of PC approxiatmion}

\item{L}{projection matrix, this can be Rx(R-1)}

\item{Q}{rank of SFA approximation}

\item{w, }{the Mx(R-1) matrix of projected strong t statistics to fit Bovy}

\item{vj}{the array of LVL' (JxR-1xR-1)}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\details{
wrapper to compute denoised estimates of the fuller rank covariance matrices

note that w and vj, L should all by R-1, while t.stat and factor mat should be R
}
