% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{deconvolution.em.with.loop}
\alias{deconvolution.em.with.loop}
\title{deconvolution.em.with.loop}
\usage{
deconvolution.em.with.loop(t.stat, factor.mat, lambda.mat, K, P,
  permsnp = 1000, maxiter = 1000)
}
\arguments{
\item{t.stat}{}

\item{P}{= rank of PC approxiatmion}

\item{permsnp}{= number of strong stats you want to train on (default is 1000)}

\item{Q}{= rank of SFA approximation}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\description{
deconvolution.em.with.loop
}
\details{
wrapper to compute denoised estimates of the fuller rank covariance matrices
}

