% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah_mixem.R
\name{dim.true.cov.fun}
\alias{dim.true.cov.fun}
\title{dim.true.cov.fun}
\usage{
\method{dim}{true.cov.fun}(max.step.unlist)
}
\details{
outputs a list of dimensions for both the true covariance matrices and the vector of pis
}
