% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixEm.R
\name{em.array.generator}
\alias{em.array.generator}
\title{em.array.generator}
\usage{
em.array.generator(b.j.hat, J, A, se.j.hat, covmat, lik.mat)
}
\arguments{
\item{b.j.hat}{= 1xR vector of MLEs}

\item{covmat}{= K dimensional list}

\item{lik.mat}{JxK matrix of likelihoods}

\item{se.j.jat=1xR}{vector of their standard errors}
}
\value{
list of JxKxR conditional posterior means, JxKxRxR covariance matries and JxK normalized likelihoods
}
\description{
em.array.generator
}

