% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewSimulations.R
\name{factor_sim_new}
\alias{factor_sim_new}
\title{factor.sim.new}
\usage{
factor_sim_new(J, d = 44, betasd = 1, esd = 0.11, tspec = 0, n = 400,
  R = 44)
}
\arguments{
\item{J}{Number of Gene-SNP Pairs and by definition, only 0.8 percent of them will be true eQTL; pi0 is 0.20 and the number of genes is J/100)}

\item{d}{Number of subgroups}

\item{betasd}{size of covariance of true effects}

\item{esd}{standard error}

\item{n}{number of real associations}

\item{Creates}{factors with magnitude based on betasd where each betaj is loaded on a minimal number of factors by simulating from rmvnorm}
}
\details{
Generate J beta,betahats, T statistics across R tissues from a set of covariance matrices

to simulate strong sharing, I use the gtex covariance matrices
}
