% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{fixpoint.cov}
\alias{fixpoint.cov}
\title{fixpoint.cov}
\usage{
fixpoint.cov(max.step.unlist, b.j.hat, se.j.hat)
}
\arguments{
\item{max.step}{a list of K pis and the true.covs arrays KxRxR which will be parsed by the e.step}

\item{dimtruecovs}{= kxrxr}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\details{
combines e and m step
}

