% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{gen.mean}
\alias{gen.mean}
\title{gen.XXX to compute total weight mean (1 x R vector) for each gene-snp pair}
\usage{
gen.mean(post.means, post.weights, A)
}
\value{
JxR matrix of grand means, grand nulls, grand lfsr
}
\description{
gen.XXX to compute total weight mean (1 x R vector) for each gene-snp pair
}

