% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{genlvlarray}
\alias{genlvlarray}
\title{genlvlarray}
\usage{
genlvlarray(s.j, L)
}
\arguments{
\item{s.j}{matrix of unscaled standard errors}

\item{L}{projection matrix, this can be Rx(R-1)}
}
\value{
the JxRxR (or R-1) array of marginal variances
}

