% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.bma.only}
\alias{get.prior.covar.bma.only}
\title{get.prior.covar.bma.only}
\usage{
get.prior.covar.bma.only(R, omega.table)
}
\arguments{
\item{R}{number of tissues}

\item{omega.table}{L vector grid weights}
}
\value{
a L x K list of covariance matrices
}
