% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.nopcs}
\alias{get.prior.covar.nopcs}
\title{get.prior.covar.nopcs}
\usage{
get.prior.covar.nopcs(lambda.mat, Q, X.c, simat, omega.table, bma = TRUE)
}
\description{
get.prior.covar.nopcs
}

