% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.Ukl}
\alias{get.prior.covar.Ukl}
\title{Get prior covariances}
\usage{
get.prior.covar.Ukl(P, L, R, factors, omega.table)
}
\arguments{
\item{P}{integer number of PCs}

\item{L}{integer number of gridweights}
}
\value{
U.0kl L dimensional list of K dimensional list with prior covairnace matrix for each grid weight, prior covariance pair
}
\description{
Compute the prior covariance matrix for each gene SNP pair in the rows and componenets in the columns
}

