% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{get.prior.covar.with.all.max.step.no.pc}
\alias{get.prior.covar.with.all.max.step.no.pc}
\title{get.prior.covar.with.all.max.step}
\usage{
get.prior.covar.with.all.max.step.no.pc(X.c, max.step, lambda.mat, Q,
  factor.mat, omega.table, bma = TRUE)
}
\arguments{
\item{Q}{rank of SFA approximation}

\item{t}{stat}

\item{P}{rank of PC approxiatmion}

\item{permsnp}{= number of strong stats you want to train on (default is 1000)}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\details{
wrapper to compute denoised estimates of the fuller rank covariance matrices
}
