% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.with.heterogeneity}
\alias{get.prior.covar.with.heterogeneity}
\title{get.prior.covar.with.heterogeneity}
\usage{
get.prior.covar.with.heterogeneity(R, omega.table)
}
\arguments{
\item{R}{number of tissues}

\item{omega.table}{L vector grid weights}
}
\value{
a L x K list of covariance matrices
}

