% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthscripts.R
\name{get.prior.covar.with.truth}
\alias{get.prior.covar.with.truth}
\title{get.prior.covar.with.truth}
\usage{
get.prior.covar.with.truth(R, omega.table, covlist)
}
\arguments{
\item{R}{number of tissues}

\item{omega.table}{L vector grid weights}
}
\value{
a L x K list of covariance matrices
}
