% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{het.norm.array.generator}
\alias{het.norm.array.generator}
\title{het.norm.array.generator}
\usage{
het.norm.array.generator(j, b.j.hat, se.j.hat, se.beta.hat, covmat, pis, sim,
  thresh)
}
\arguments{
\item{b.j.hat}{= 1xR vector of MLEs}

\item{covmat}{K list of prior covariance matrices}

\item{sim}{number of simulations per gene snp pair}

\item{se.j.hat=1xR}{vector of their standard errors}

\item{se.beta.hat=1xR}{vector of standard errors of betas for conversion}

\item{pi}{K vector of prior weights}
}
\value{
Proportion of Time signs disagree
}
\details{
simulates from multivariate normal according to componenet responsibility; counts number of times there exist discordant signs among the simulations for each gene snp pair
}
