% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{heter.mixture}
\alias{heter.mixture}
\title{heter.mixture}
\usage{
heter.mixture(lfsr.mat, thresh, posterior.means)
}
\arguments{
\item{lfsr}{= Jx(R+1)}

\item{se.j.hat=1xR}{vector of their standard errors}

\item{max.step=list}{containing two arrays: true.covs= KxRxR arrays of prior covariance matrices}

\item{pi}{K vector of prior weights}
}
\value{
list of 3 arrays: JxKxR conditional posterior means, JxKxRxR posterior covariance matries and JxK normalized likelihoods (P(K|Data))
}

