% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{hm.weight.gen}
\alias{hm.weight.gen}
\title{hm.weight.gen}
\usage{
hm.weight.gen(lik.mat, jtrain, jtest)
}
\arguments{
\item{lik.mat}{= JxK matrix of marginal likelihoods}

\item{Jtrain}{= size of train set}

\item{Jtest=size}{of test set}
}
\value{
compute HM weights on training and test set for compoarsion
}
\description{
hm.weight.gen
}

