% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NewSimulations.R
\name{independent.from.omega}
\alias{independent.from.omega}
\title{independent.from.omega}
\usage{
independent.from.omega(J, d = 44, betasd = 1, esd = 0.1, tspec = 0)
}
\description{
independent.from.omega
}
\details{
simulate shared effects with randomly chosen omegas
}

