% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{init.covmat.single.with.projection}
\alias{init.covmat.single.with.projection}
\title{init.covmat.with.projection}
\usage{
init.covmat.single.with.projection(t.stat = t.stat, factor.mat = factors,
  lambda.mat = lambda, P, Q)
}
\arguments{
\item{t.stat}{matrix of transformed strong statistics (MxR) from which to derive covariance matrices (note this P will need to be RxR)}

\item{factor.mat}{KxR matrix of factors that have computed on strong projected ts from SFA}

\item{lambda.mat}{KxR matric of loadings that have been computed strong projected ts , also from SFA}

\item{P}{rank of SVD approximation}

\item{K}{number of components to fit (i.e., the first dimension of the array), default is 3+Q}
}
\value{
KxRxR matric of prior covariance matrices to initialize the EM
}

