% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{lik.func.em}
\alias{lik.func.em}
\title{lik.func.em}
\usage{
lik.func.em(true.covs, b.mle, V.j.hat)
}
\arguments{
\item{true.covs}{= KxRxR array of prior covariance matrices}

\item{b.mle=Rx1}{vector of mles}

\item{B.j.hat}{= matrix of squared standard erros}
}
\value{
1xK vector of likelihoods for the jth gene snp pair across all K covariance matrices
}
\description{
lik.func.em
}

