% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{lik.func}
\alias{lik.func}
\title{lik.func}
\usage{
lik.func(b.mle, V.gp.hat, covmat)
}
\arguments{
\item{b.mle}{Rx1 vector of mles}

\item{V.gp.hat}{RxR matrix of standard errors}

\item{U.0kl}{L dimensional list of K dimensional list with prior covairnace matrix for each grid weight, prior covariance pair}
}
\description{
lik.func
}
\details{
computes likelihood for each betahat
}

