% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lik.func.with.tol}
\alias{lik.func.with.tol}
\title{lik.func.with.tol}
\usage{
lik.func.with.tol(b.mle, R, V.gp.hat, covmat)
}
\arguments{
\item{b.mle}{Rx1 vector of mles}

\item{V.gp.hat}{RxR matrix of standard errors}

\item{U.0kl}{L dimensional list of K dimensional list with prior covairnace matrix for each grid weight, prior covariance pair}
}
\details{
computes likelihood for each betahat
}
