% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{log.lik.func}
\alias{log.lik.func}
\title{log.lik.func}
\usage{
\method{log}{lik.func}(b.mle, V.gp.hat, covmat)
}
\arguments{
\item{b.mle}{Rx1 vector of mles}

\item{V.gp.hat}{RxR matrix of standard errors}

\item{U.0kl}{L dimensional list of K dimensional list with prior covairnace matrix for each grid weight, prior covariance pair}
}
\details{
computes log.likelihood for each betahat
}

