% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{max.step.func}
\alias{max.step.func}
\title{max.step.func}
\usage{
\method{max}{step.func}(post.means, post.covs, q.mat)
}
\arguments{
\item{post.means}{JxKxR matrix of posterior means}

\item{post.covs}{JxKxR matrix of posterior covariances}

\item{q.mat}{JxK matrix of normalixed likelihoods e.g., p(K|D)}
}
\value{
K x R xR array of 'true covariance matrices' and Kx1 vector of pis
}
\description{
max.step.func
}

