% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{mult.tissue.grid}
\alias{mult.tissue.grid}
\title{mult.tissue.grid}
\usage{
mult.tissue.grid(mult, betahat, sebetahat)
}
\arguments{
\item{mult}{is density}

\item{betahat}{is matrix of betahats}

\item{sebetahat}{is matrix of sebetahats}
}
\value{
omega
}
\description{
mult.tissue.grid
}

