% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testscript.R
\name{mult.tissue.omega}
\alias{mult.tissue.omega}
\title{mult.tissue.omega}
\usage{
mult.tissue.omega(mult, betahat, sebetahat, minparam, multone, maxparam)
}
\arguments{
\item{betahat}{= JxR matrix of betahats (or zs)}

\item{minparam}{= what the standard error is divided by (want to be small)}

\item{multone}{= density parameter (intervals between points for within tissue variances, default 2 so that the ratio between sd is sqrt(2))}

\item{maxparam}{= how big the grid needs to be}

\item{sebetahat=}{JxR matrix of standard errs}
}
\value{
vector of variances for all tissues
}
\concept{
this computes a rough largest value you'd want to use, based on idea that sigmaamax^2 + sebetahat^2 should be at least betahat^2
}
