% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{post.array.generator}
\alias{post.array.generator}
\title{creat an array of posterior quantities}
\usage{
post.array.generator(b.gp.hat, se.gp.hat, covmat)
}
\arguments{
\item{covmat}{= LxK dimenstional (unlisted list) of prior covariance matrices}

\item{b.mle}{=  Rx1 vector of beta.hats}

\item{V.gp.hat}{= Rx1 vector of standard errors}
}
\value{
post.means JxKxR array of posterior means, correspodning to the posterior mean ##' for the Jth individual in the Kth compoenent across all R tissues

post.covs JxKxR array of posterior vars, correspodning to the posterior vars ##' for the Jth individual in the Kth compoenent across all R tissues

post.ups JxKxR array of posterior tailprobs, corresponding to the marginal
upper tail probability for the Jth individual in the Kth compoenent across all R

post.ups JxKxR array of posterior tailprobs, corresponding to the marginal
upper  tail probability for the Jth individual in the Kth component across all R

post.nulls JxKxR array of posterior nullprobs, corresponding to the marginal
 "null probability"" for the Jth individual in the Kth component across all R
}
\description{
creat an array of posterior quantities
}

