% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{post.array.per.snp.no.baseline}
\alias{post.array.per.snp.no.baseline}
\title{post.array.per.snp.no.baseline}
\usage{
post.array.per.snp.no.baseline(j, covmat, b.gp.hat, se.gp.hat, L)
}
\arguments{
\item{b.gp.hat}{projected L}

\item{se.gp.hat}{standard errors of unprojected estimates}

\item{L}{projection matrix}
}
