% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{post.array.per.snp.with.lvlmat}
\alias{post.array.per.snp.with.lvlmat}
\title{post.array.per.snp.with.lvlmat}
\usage{
post.array.per.snp.with.lvlmat(j, covmat, b.gp.hat, lvlmat)
}
\arguments{
\item{cormat}{is the matrix of corerlation of errors}

\item{segphat}{is a the JxR matrix of standard errors for scaling}
}
