% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{post.array.per.snp.with.mat}
\alias{post.array.per.snp.with.mat}
\title{post.array.per.snp.with.mat}
\usage{
post.array.per.snp.with.mat(j, covmat, b.gp.hat, se.mat)
}
\description{
post.array.per.snp.with.mat
}

