% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{post.array.per.snp.with.vmat}
\alias{post.array.per.snp.with.vmat}
\title{post.array.per.snp.with.vmat}
\usage{
post.array.per.snp.with.vmat(j, covmat, b.gp.hat, se.gp.hat, cormat)
}
\arguments{
\item{cormat}{is the matrix of corerlation of errors}

\item{segphat}{is a the JxR matrix of standard errors for scaling}
}

