% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{post.b.gpkl.cov}
\alias{post.b.gpkl.cov}
\title{post.b.gpkl.cov}
\usage{
post.b.gpkl.cov(V.gp.hat.inv, U.0k.l)
}
\arguments{
\item{U.0k.l}{let U.0k.l represent a specific matrix in U.0kl (.e.g, U.0kl[[l]][[k]])}
}
\value{
post.b.gpkl.cov ## returns an R*R posterior covariance matrix for jth gene snp pair

post.b.gpkl.mean return a 1 * R vector of posterior means for a given prior covariance matrix
}

