% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{post.b.gpkl.mean}
\alias{post.b.gpkl.mean}
\title{post.b.gpkl.mean}
\usage{
post.b.gpkl.mean(b.mle, V.gp.hat.inv, U.gp1kl)
}
\arguments{
\item{U.0k.l}{let U.0k.l represent a specific matrix in U.0kl (.e.g, U.0kl[[l]][[k]])}
}
\value{
post.b.gpkl.mean return a 1 * R vector of posterior means for a given prior covariance matrix
}
\description{
post.b.gpkl.mean
}

