% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixEm.R
\name{post.b.jk.cov}
\alias{post.b.jk.cov}
\title{post.b.jk.cov}
\usage{
post.b.jk.cov(V.j.hat.inv, U.0k)
}
\value{
RxR posterior covariance matrix for a given prior covariance matrix
}
\description{
post.b.jk.cov
}

