% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{post.b.jk.ed.cov}
\alias{post.b.jk.ed.cov}
\title{post.b.jk.ed.cov}
\usage{
post.b.jk.ed.cov(tinv, U.k)
}
\arguments{
\item{tinv}{= variance of marginal distirbution of bhat, RxR matrix (e.g.,tinv=solve(U.k+V.j.hat)) )}

\item{U.k}{= prior covariance matrix (RxR)}
}
\value{
RxR posterior covariance matrix for the Jth gene snp pair at the kth componenet
}

