% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{post.b.jk.ed.mean}
\alias{post.b.jk.ed.mean}
\title{post.b.jk.ed.mean}
\usage{
post.b.jk.ed.mean(b.mle, tinv, U.k)
}
\arguments{
\item{tinv}{= variance of marginal distirbution of bhat, RxR matrix (e.g.,tinv=solve(U.k+V.j.hat)) )}

\item{U.k}{= RxR prior covariance matrix}

\item{b.mle=Rx1}{vector of mles}
}
\value{
component specific vector of posterior means for the jth gene snp pair
}

