% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixEm.R
\name{post.b.jk.mean}
\alias{post.b.jk.mean}
\title{post.b.jk.mean}
\usage{
post.b.jk.mean(b.mle, V.j.hat.inv, U.1jk)
}
\arguments{
\item{U.0k.l}{let U.0k.l represent a specific matrix in U.0kl (.e.g, U.0kl[[l]][[k]])}
}
\value{
return a 1 * R vector of posterior means for a given prior covariance matrix
}
\description{
post.b.jk.mean
}

