% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{post.cov.with.proj}
\alias{post.cov.with.proj}
\title{post.cov.with.proj}
\usage{
post.cov.with.proj(tinv, U.k, L)
}
\arguments{
\item{tinv}{inverse of marginal covariance of b.mle}

\item{U.k}{prior covariance of true effect}

\item{L}{projection matrix}

\item{b.gp.hat}{projected L}
}

