% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{post.mean.with.proj}
\alias{post.mean.with.proj}
\title{post.mean.with.proj}
\usage{
post.mean.with.proj(b.mle, tinv, U.k, L)
}
\arguments{
\item{b.mle}{projected L}

\item{tinv}{inverse of marginal covariance of b.mle}

\item{U.k}{prior covariance of true effect}

\item{L}{projection matrix}
}

