% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim.array.generation}
\alias{sim.array.generation}
\title{sim.array.generator}
\usage{
sim.array.generation(j, b.j.hat, se.j.hat, covmat, pis, sim)
}
\arguments{
\item{b.j.hat}{= 1xR vector of MLEs}

\item{covmat}{K list of prior covariance matrices}

\item{sim}{number of simulations per gene snp pair}

\item{se.j.hat=1xR}{vector of their standard errors}

\item{pi}{K vector of prior weights}
}
\value{
Proportion of Time signs disagree
}
\details{
simulates from multivariate normal according to componenet responsibility; counts number of times there exist discordant signs among the simulations for each gene snp pair
}
