% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{test.funct}
\alias{test.funct}
\title{test.funct}
\usage{
test.funct(j, max.step.unlist, k, R)
}
\arguments{
\item{j}{gene snp pair of interest}

\item{k}{componenet of interest}

\item{R}{number of tissues}
}
\details{
shows that the e step properly stores the posterior mean for a given gene snp pair and componenet combination
}
