% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{total.covs.partone.persnp}
\alias{total.covs.partone.persnp}
\title{total.covs.partone.persnp}
\usage{
total.covs.partone.persnp(post.means, post.covs, post.weights)
}
\description{
total.covs.partone.persnp
}

