% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total.covs.partone}
\alias{total.covs.partone}
\title{total.covs.partone}
\usage{
total.covs.partone(j, post.means, post.covs, post.weights)
}
\arguments{
\item{post.covs}{J x K x R arrays of posterior marginal variance (diagonals of posteior covmat) for each snp in each component in each tissue}

\item{post.weights}{J x K matrix of posterior weights for each componenet and for each snp}
}
\description{
generate a K x R matrix of post.weighted marginal variances and sum them to get total weighted
}
