% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total.down}
\alias{total.down}
\title{total.down}
\usage{
total.down(j, post.downs, post.weights)
}
\arguments{
\item{post.downs}{J x K x R arrays of posterior lower tail probabilities for each snp in each component in each tissue}

\item{post.weights}{J x K matrix of posterior weights for each componenet and for each snp}
}
\description{
generate a K x R matrix of post.weighted lower tail probabilites and sum them to get total weighted
}
