% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total.lik.func}
\alias{total.lik.func}
\title{total.lik.func}
\usage{
total.lik.func(test, pis)
}
\arguments{
\item{test}{= J x R matrix of likielihoods from test data set}

\item{pis}{HM weights}
}
\value{
likelihood of whole data set
}
