% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total.null}
\alias{total.null}
\title{total.null}
\usage{
total.null(j, post.nulls, post.weights)
}
\arguments{
\item{post.nulls}{J x K x R arrays of posterior upper tail probabilities for each snp in each component in each tissue}

\item{post.weights}{J x K matrix of posterior weights for each componenet and for each snp}
}
\description{
generate a K x R matrix of post.weighted null probabilites and sum them to get total weighted
}
