% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mashnobasescripts.R
\name{total.quant.per.snp.no.baseline}
\alias{total.quant.per.snp.no.baseline}
\title{total.quant.per.snp.nobaseline}
\usage{
total.quant.per.snp.no.baseline(j, covmat, b.gp.hat, se.gp.hat, pis, A,
  checkpoint = FALSE, L = diag(1, d))
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles (let's use these as the transformed, consistnent with w = LChat, so centered)}

\item{se.gp.hat}{JxR matrix of standard errors of uncentered estimates}

\item{pis}{Kx1 matrix of prior weights}

\item{L}{projection matrix: default IDentity; can be the centering matrix that was used to center uncentered averages}

\item{column}{dimension of b.gp.hat (in the untransformed case, this is the same as the dimension of the true effects, but not in transformed case)}
}
\value{
writes the posterior weighted quantities to a file
}

