% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total.quant.per.snp.with.lvlarray}
\alias{total.quant.per.snp.with.lvlarray}
\title{total.quant.per.snp.with.lvlarray}
\usage{
total.quant.per.snp.with.lvlarray(j, covmat, b.gp.hat, lvlarray, pis, A,
  checkpoint = FALSE)
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles}

\item{pis}{Kx1 matrix of prior weights}

\item{se.gp.hat}{JxR matrix of standarderrors}

\item{cormat}{RxR matrix of variance}
}
\value{
writes the posterior weighted quantities to a file
}
