% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{total.quant.per.snp.with.var}
\alias{total.quant.per.snp.with.var}
\title{total.quant.per.snp.with.var}
\usage{
total.quant.per.snp.with.var(j, covmat, b.gp.hat, se.mat, pis, A,
  checkpoint = FALSE)
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles}

\item{pis}{Kx1 matrix of prior weights}

\item{se.gp.hat}{JxR matrix of standard errs}
}
\value{
writes the posterior weighted quantities to a file
}
\description{
total.quant.per.snp.with.var
}

