% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{total.up}
\alias{total.up}
\title{total.up}
\usage{
total.up(j, post.ups, post.weights)
}
\arguments{
\item{post.ups}{J x K x R arrays of posterior upper tail probabilities for each snp in each component in each tissue}

\item{post.weights}{J x K matrix of posterior weights for each componenet and for each snp}
}
\value{
get vector of total weighted quanties for each gene SNP pair
}
\description{
generate a K x R matrix of post.weighted upper tail probabilites and sum them to get total weighted
}

