% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{autoselect.mix.sd}
\alias{autoselect.mix.sd}
\title{Select a grid}
\usage{
autoselect.mix.sd(betahat, sebetahat, mult)
}
\arguments{
\item{sebetahat}{is a J x R matrix of their standard errors (or 1s, if T staitsitcs)}

\item{beta}{hat is an J x R matrix of betahats (or t statistics)}
}
\value{
omega, a list of stretch factors by which to scale each covariance matrix U_k
}
\description{
mult.tissue.grid computes a set of omega 'stretch' factors in log 2 multiples of sd; the max and min for each tissue are chosen by autoselect.mix.sd
}

