% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compare.func}
\alias{compare.func}
\title{compare.func}
\usage{
compare.func(j, all.upper, all.lower)
}
\arguments{
\item{all.upper}{and all.lower are JxR matrices of weighted upper and lower tail probabilities for all gene pairs}
}
\value{
1 x R vector of lfsr at each tissue
}
\details{
compares upper and lower tail probability at each tissue to determine larger
}
