% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.covmat}
\alias{compute.covmat}
\title{compute.covmat}
\usage{
compute.covmat(b.gp.hat, sebetahat, Q, X.c, lambda.mat, P, A, factor.mat,
  bma = TRUE)
}
\value{
A list of covariance matrices
}
\description{
compute.covmat
}

