% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.covmat}
\alias{compute.covmat}
\title{compute.covmat}
\usage{
compute.covmat(b.gp.hat, sebetahat, Q, t.stat, lambda.mat, P, A, factor.mat,
  bma = TRUE, zero = FALSE, maxp = 1, minp = 1, power = 1)
}
\arguments{
\item{b.gp.hat}{a JxR matrix of betahats - necessary for grid computation}

\item{sebetahat}{a JxR matrix of their standard errors - necessary for grid computation}

\item{Q}{number of factors to consider}

\item{t.stat}{matrix of strong t statistics (needs to be R in columns but not necessarily j in rows)}

\item{factor}{mat a fxR matrix of factors}

\item{BMA=true}{wheter or not to use singleton and shared config}

\item{zero=true}{whether or not to include zero matrix}
}
\value{
A list of covariance matrices
}

