% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{compute.hm.covmat.all.max.step}
\alias{compute.hm.covmat.all.max.step}
\title{compute.hm.covmat.all.max.step}
\usage{
compute.hm.covmat.all.max.step(b.hat, se.hat, t.stat, Q, lambda.mat, A,
  factor.mat, max.step, zero = FALSE, maxp = 1, minp = 1, power = 1,
  bma = TRUE)
}
\arguments{
\item{b.hat, se.hat}{matrices to help choose grid weights (JxR)}

\item{Q}{number of single rank factors to include in the set of covariance matrices}

\item{maxp, }{minp - what should be the maximum (minimum) factor by which minimum sd and maximum autoselected sd are stretched, default=1}

\item{max.step:}{the list ouput of deconvolution EM (list of length 2, contains the denoised covariance  matrices and vector of pi)}

\item{P}{PC approximation}

\item{v.j}{a JxR matrix of standard errors for scaling (in the zstat case this is all 1s)}

\item{BMA}{Whether or not to include singleton and full configurations}

\item{Power}{should omega be autoselect.sd or autoselect.sd ^2; recall that in gtex, this is 1}
}
\value{
a list of covariance matrices
}
\details{
use a prespecified X`X from the output of the 'deconvoluting EM'
}

