% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{compute.hm.train.log.lik.vmat}
\alias{compute.hm.train.log.lik.vmat}
\title{compute.hm.train.log.lik.vmat}
\usage{
compute.hm.train.log.lik.vmat(train.b, v.mat, covmat, A)
}
\description{
compute.hm.train.log.lik.vmat
}
\details{
here, use the matrix of residual variation that may have non-diagonal elements
}

