% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{deconvolution.em.with.bovy}
\alias{deconvolution.em.with.bovy}
\title{deconvolution.em.bovy}
\usage{
deconvolution.em.with.bovy(t.stat, factor.mat, v.j, lambda.mat, K, P)
}
\arguments{
\item{t.stat}{}

\item{P}{= rank of PC approxiatmion}

\item{Q}{= rank of SFA approximation}

\item{permsnp}{= number of strong stats you want to train on (default is 1000)}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\description{
deconvolution.em.bovy
}
\details{
wrapper to compute denoised estimates of the fuller rank covariance matrices
}

