% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah_mixem.R
\name{deconvolution.em.with.max.iter}
\alias{deconvolution.em.with.max.iter}
\title{deconvolution.em.with.max.iter}
\usage{
deconvolution.em.with.max.iter(t.stat, factor.mat, lambda.mat, K, P,
  permsnp = 1000, maxiter = 1000)
}
\arguments{
\item{P}{rank of PC approxiatmion}

\item{permsnp}{= number of strong stats you want to train on (default is 1000)}

\item{t}{stat}

\item{Q}{rank of SFA approximation}
}
\value{
a 2 element list of K pis and the KxRxR true.covariance arrays
}
\details{
wrapper to compute denoised estimates of the fuller rank covariance matrices
}
