% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarah's_mixem.R
\name{em.array.generator}
\alias{em.array.generator}
\title{em.array.generator}
\usage{
em.array.generator(max.step, b.j.hat, se.j.hat)
}
\arguments{
\item{b.j.hat}{= 1xR vector of MLEs}

\item{se.j.hat=1xR}{vector of their standard errors}

\item{max.step=list}{containing two arrays: true.covs= KxRxR arrays of prior covariance matrices}

\item{pi}{K vector of prior weights}
}
\value{
list of 3 arrays: JxKxR conditional posterior means, JxKxRxR posterior covariance matries and JxK normalized likelihoods (P(K|Data))
}
