% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{fixpoint.play}
\alias{fixpoint.play}
\title{fixpoint.play}
\usage{
fixpoint.play(pi, matrix_lik, prior)
}
\value{
pi
}
\description{
fixpoint.play
}
\details{
returns pis for squareEM recognizing that if the rowSums are 0 for certain likmat entries, these need to be removed
}

