% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.Ukl}
\alias{get.prior.covar.Ukl}
\title{get.prior.covar.Ukl}
\usage{
get.prior.covar.Ukl(P, lambda.mat, Q, factor.mat, omega.table, bma = TRUE)
}
\arguments{
\item{P}{number of pcs}

\item{lambda.mat}{matrix of lambdas}

\item{Q}{number factor}

\item{factor.mat}{}

\item{omega.table}{}

\item{bma=TRUE}{}
}
\value{
U.0kl
}
\description{
get.prior.covar.Ukl
}

