% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.Ukl}
\alias{get.prior.covar.Ukl}
\title{get.prior.covar.Ukl}
\usage{
get.prior.covar.Ukl(P, X.c, lambda.mat, Q, factor.mat, omega.table,
  bma = TRUE)
}
\arguments{
\item{P}{number of pcs (must be greater than 1)}

\item{X.c}{= column centered matrix of strong t statistics from which to derive covariance matrices}

\item{lambda.mat}{matrix of JxQ lambdas}

\item{Q}{number of single rank factor matrices (if 0, then only the rank 5 approximation wil be used)}

\item{factor.mat}{nfactors x R tissue matrix of factors}

\item{omega.table}{Lx2 table of grid weights}

\item{bma}{whether or not to use singleton and shared configurations}
}
\value{
U.0kl
}
\description{
get.prior.covar.Ukl
}

