% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.prior.covar.Ukl.with.rho}
\alias{get.prior.covar.Ukl.with.rho}
\title{get.prior.covar.Ukl.with.rho}
\usage{
get.prior.covar.Ukl.with.rho(P, lambda.mat, X.c, Q, factor.mat, omega.table,
  rhos)
}
\arguments{
\item{P}{integer number of PCs}

\item{X.c}{= column centered matrix of strong t statistics from which to derive covariance matrices}

\item{omega.table}{table of grid weights}

\item{rhos}{correlation factor for morphed matrices}
}
\value{
U.0kl L dimensional list of K dimensional list with prior covairnace matrix for each grid weight, prior covariance pair
}
\description{
Compute the prior covariance matrix for each gene SNP pair in the rows and componenets in the columns
}
