% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mult.tissue.grid}
\alias{mult.tissue.grid}
\title{mult.tissue.grid}
\usage{
mult.tissue.grid(mult = sqrt(2), betahat, sebetahat, maxp, minp)
}
\arguments{
\item{mult}{is density}

\item{betahat}{is matrix of betahats}

\item{sebetahat}{is matrix of sebetahats}
}
\value{
omega
}
