% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{total.quant.per.snp.with.vmat}
\alias{total.quant.per.snp.with.vmat}
\title{total.quant.per.snp.with.vmat}
\usage{
total.quant.per.snp.with.vmat(j, covmat, b.gp.hat, cormat, se.gp.hat, pis, A,
  checkpoint = FALSE)
}
\arguments{
\item{covmat}{a K list of covariance matrices}

\item{b.gp.hat}{JxR matrix of mles}

\item{cormat}{RxR matrix of variance}

\item{se.gp.hat}{JxR matrix of standarderrors}

\item{pis}{Kx1 matrix of prior weights}
}
\value{
writes the posterior weighted quantities to a file
}
